separate out SVGP-specific tests from test_methods into test_svgp
